Add beginning functions for calculator
I stored the various operators into one object based on a eariler
practice problem I work through. I think it may be easier to do it this
way because of trying to split the string (equation) into an array, and
basing which operation to perform of of array[1]. However later I may
come back to change this to individual functions for each operator.